refactor errors

+++ src/main/java/com/coreone/back/service/SectorService.java
@@ -5,6 +5,7 @@ import com.coreone.back.dto.sector.CreateSectorRequestDTO;
 import com.coreone.back.dto.sector.CreateSectorResponseDTO;
 import com.coreone.back.dto.sector.GetSectorResponse;
 import com.coreone.back.errors.NotFoundException;
+import com.coreone.back.errors.UnauthorizedException;
 import com.coreone.back.errors.sectorlAlreadyExistsException;
 import com.coreone.back.mapper.SectorMapper;
 import com.coreone.back.repository.SectorRepository;
@@ -51,7 +52,10 @@ public class SectorService {
     }
 
     public Page<GetSectorResponse> listAllEnterpriseSectors(UUID id, int page, int size) {
-        var enterprise = enterpriseService.findById(id);
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+
+        var enterprise = enterpriseService.findById(user.getEnterprise().getId());
 
         Pageable pageable = PageRequest.of(page, size);
 
@@ -68,7 +72,14 @@ public class SectorService {
         );
     }
 
-    public String delete(UUID id) {
+    public String delete(UUID id, UUID userId) {
+        var user = userRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+
+        if (!user.getEnterprise().getId().equals(userId)) {
+            throw new UnauthorizedException("Unauthorized request");
+        }
+
         var sector = findById(id);
 
         repository.delete(sector);
